Add max product version option
This replicates some of the MSI upgrade logic.
Allows only products with versions below a certain
number to be uninstalled.

For example, this allows upgrading from a per-user
installation with version 1 to a per-machine installation
with version 2.

Otherwise, the uninstall utility will remove the version 2
without discrimination during a repair.